+++ next.config.js
@@ -12,6 +12,11 @@ module.exports = {
         destination: '/dashboard',
         permanent: true,
       },
+      {
+        source: '/login',
+        destination: '/login/nl',
+        permanent: true,
+      },
     ];
   },
 };
