+++ src/components/LoginButton.js
@@ -1,9 +1,12 @@
 import PropTypes from 'prop-types';
 import { useCookies } from 'react-cookie';
+import { useTranslation } from 'react-i18next';
 import { Button } from './publiq-ui/Button';
 
-const LoginButton = ({ language, className }) => {
+const LoginButton = ({ className }) => {
   const [, , removeCookie] = useCookies(['token', 'user']);
+
+  const { i18n } = useTranslation();
 
   const buildBaseUrl = () =>
     `${window.location.protocol}//${window.location.host}`;
@@ -18,7 +21,7 @@ const LoginButton = ({ language, className }) => {
 
     const queryString = new URLSearchParams({
       destination: buildBaseUrl(),
-      lang: language,
+      lang: i18n.language,
     }).toString();
 
     window.location.href = `${process.env.NEXT_PUBLIC_AUTH_URL}/connect?${queryString}`;
@@ -32,7 +35,6 @@ const LoginButton = ({ language, className }) => {
 };
 
 LoginButton.propTypes = {
-  language: PropTypes.oneOf(['nl', 'fr']).isRequired,
   className: PropTypes.string,
 };
