+++ src/pages/[...params].js
@@ -1,12 +1,13 @@
 import { useRouter } from 'next/router';
 import { useCookies } from 'react-cookie';
 import i18next from 'i18next';
-import { useEffect } from 'react';
+import { useEffect, useState } from 'react';
+import { Box } from '../components/publiq-ui/Box';
 
 const Fallback = () => {
-  let iframe;
   const router = useRouter();
   const [cookies] = useCookies(['token']);
+  const [legacyPath, setLegacyPath] = useState('');
 
   useEffect(() => {
     if (!cookies.token) {
@@ -19,24 +20,24 @@ const Fallback = () => {
       lang: i18next.language,
     }).toString();
 
+
     const path = router.asPath ? router.asPath : '';
     const parsedQueryString = queryString ? `?${queryString}` : '';
-    const generatedPath = `${process.env.NEXT_PUBLIC_LEGACY_APP_URL}${path}${parsedQueryString}`;
-    iframe = `<iframe src="${generatedPath}" />`;
+    setLegacyPath(
+      `${process.env.NEXT_PUBLIC_LEGACY_APP_URL}${path}${parsedQueryString}`,
+    );
   }, []);
 
-  if (!iframe) return null;
+  if (!legacyPath) return null;
 
   return (
-    <div
+    <Box
+      as="iframe"
+      src={legacyPath}
       css={`
-        iframe {
-          height: 100vh;
-          width: 100%;
-          border: 0;
-        }
+        height: 100vh;
+        width: 100%;
       `}
-      dangerouslySetInnerHTML={{ __html: iframe }}
     />
   );
 };
